+++ packages/shared/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 6.2.2-dev.61 (2026-08-26)
+
+This was a version bump only for @powerhousedao/shared to align it with other projects, there were no code changes.
+
+## 6.2.2-dev.60 (2026-08-26)
+
+This was a version bump only for @powerhousedao/shared to align it with other projects, there were no code changes.
+
 ## 6.2.2-dev.59 (2026-08-25)
 
 ### 🩹 Fixes